feat(cloud-crawler): Create CrawlRunner class to manage crawling during a run
At present, this class manages logging and error handling for crawling during a run. In future, it can also manage adding known urls to scan results.

Note: In a few places, the use of `inversify.Container` was changed to `inversify.interfaces.Container` because functions which create the crawler only have the interface type available.