chore: ensure cli build fails if dts-bundle-generator fails (#1794)
#### Details

This PR fixes an issue with the cli package's build process where `dts-bundle-generator` failures are silently ignored by the build. Since it rewrites a published file in place, this results in publishing an incorrect version of `index.d.ts` if a failure occurs in that tool.

##### Motivation

Avoid silently publishing incorrect build output.

##### Context

This PR does not fix the root cause of why dts-bundle-generator is failing in recent builds; it just makes the failure non-silent. @DaveTryon will be looking into the current failure separately.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->

- [n/a] Addresses an existing issue: Fixes #0000
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [ ] Validated in an Azure resource group