fix: Update auth flow to handle KMSI page (#2295)
#### Details

This adds a check for the KMSI (Keep Me Signed In) page that shows up
conditionally based on a machine learning algorithm that detects high
risk logins. It also bumps the scan package version in preparation for
release.

More information on the KMSI page can be found here:
https://learn.microsoft.com/en-us/azure/active-directory/fundamentals/customize-branding#learn-about-the-stay-signed-in-prompt

##### Motivation

Fix for auth

##### Context

My initial implementation did not use the try/catch block, but instead
was performing a `page.$eval` and looking for the selector. I changed it
to a `waitForSelector` with a 5000ms timeout to ensure that we aren't
missing the dialog due to a possible race condition where the page
doesn't fully render before the `page.$eval` executes.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [n/a] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [ ] Validated in an Azure resource group