Ensure that the most precise error information is included in a scan run result (#2296)
#### Details
A previous change serialized a BrowserError to string before passing it
into Runner.setRunResult. This had the unintended side effect of
resulting in all BrowserErrors being reported as InternalError, instead
of being mapped to the corresponding ScanRunErrorCode.

This change explicitly converts Error and BrowserError objects into
ScanError objects before passing them to setRunResult. This works
because both BrowserError and ScanError objects have an errorType member
and the BrowserErrorTypes and ScanErrorTypes align. Error has no overlap
with ScanError so it is mapped to InternalError - which is what the
conversion handling would ultimately do with a string anyway.

For reference see also ScanResponseConverter.createFailedScanResponse
and ScanErrorConverter.getScanRunErrorCode; this is the code path
whereby a string | ScanError is converted into a ScanRunError for
inclusion in a ScanResultResponse.

##### Motivation
Addresses internal bug 2007185 written against WCP.

##### Context
This was working before by a lucky loophole. We were passing a variable
of type string | BrowserError into a type of string | ScanError. The
string types on both sides allowed it to pass compile but then sneak
through a BrowserError at runtime, which happened to be assignable and
consumable as a ScanError.

I don't love this solution but it does make the luck and assumptions
explicit, and it is narrowly scoped. Any change to the actual types
defined for OnDemandPageScanResult or ScanResultResponse would have had
a much wider ripple effect including the potential to affect a11y as
well.

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [x] Addresses an existing issue: Fixes #0000
- [n/a] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [ ] Ran precheckin (`yarn precheckin`)
- [x] Validated in an Azure resource group

Co-authored-by: pownkel <55459788+pownkel@users.noreply.github.com>