Enable crawling for URLs having hash fragments for ai-scan cli (#2524)
#### Details

Added new optional parameter keepUrlFragment to enable crawling for URLs
having hash fragments for ai-scan cli and crawler packages.
The default value of it is false so that existing scan work as is.

##### Motivation

[Feature
2114928](https://dev.azure.com/mseng/1ES/_workitems/edit/2114928): ADO
Extension should be able to crawl hash-routing SPA apps

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->

- [ ] Addresses an existing issue: Fixes #0000
- [x] Added relevant unit test for your changes. (`yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report
at: `<rootDir>/test-results/unit/coverage`
- [x] Ran precheckin (`yarn precheckin`)
- [n/a] Validated in an Azure resource group